Исправил замечания по дз №3
