+++ index.php
@@ -1,17 +1,19 @@
 <?php
 
-require_once('templates/functions.php');
-require_once('templates/data.php');
+require_once('classes/functions.php');
+require_once('classes/data.php');
 
 $page_content = include_template('main.php', [
     'categories' => $categories,
-    'ads' => $ads
+    'lots' => $lots
 ]);
 
 $layout_content = include_template('layout.php', [
     'content' => $page_content,
     'categories' => $categories,
-    'title' => 'YetiCave - Главная страница'
+    'title' => 'YetiCave - Главная страница',
+    'user_name' => $user_name,
+    'is_auth' => $is_auth
 ]);
 
 print($layout_content);
