--- templates/functions.php
@@ -1,43 +1,0 @@
-<?php
-
-function include_template($name, array $data = []) {
-    $name = 'templates/' . $name;
-    $result = '';
-
-    if (!is_readable($name)) {
-        return $result;
-    }
-
-    ob_start();
-    extract($data);
-    require $name;
-
-    $result = ob_get_clean();
-
-    return $result;
-}
-
-function esc($str) {
-    $text = htmlspecialchars($str);
-    //$text = strip_tags($str);
-
-    return $text;
-}
-
-function user_bet($bet) {
-    $html = "<b class=\"rub\"> Р</b>";
-    if ($bet >= 1000) {
-        return number_format(ceil($bet), 0, ',', " ") . $html;
-    }
-    return ceil($bet) . $html;
-}
-
-function show_date($timestamp){
-    $dt = date_create();
-    $dt = date_timestamp_set($dt, $timestamp);
-
-    $format = date_format($dt, "d.m.Y H:i");
-
-    return $format;
-}
-
