+++ templates/layout.php
@@ -26,7 +26,7 @@
                 <!-- здесь должен быть PHP код для показа меню и данных пользователя -->
                 <?php if ($is_auth == 1): ?>
                     <div class="user-menu__logged">
-                        <p><?= $user_name ?></p>
+                        <p><?= esc($user_name); ?></p>
                         <a class="user-menu__bets" href="pages/my-bets.html">Мои ставки</a>
                         <a class="user-menu__logout" href="#">Выход</a>
                     </div>
@@ -53,8 +53,8 @@
         <ul class="nav__list container">
             <!--заполните этот список из массива категорий-->
             <?php foreach ($categories as $key => $val): ?>
-                <li class="promo__item promo__item--<?= htmlspecialchars($key) ?>">
-                    <a class="promo__link" href="pages/all-lots.html"><?= htmlspecialchars($val) ?></a>
+                <li class="promo__item promo__item--<?= esc($key) ?>">
+                    <a class="promo__link" href="pages/all-lots.html"><?= esc($val) ?></a>
                 </li>
             <?php endforeach; ?>
         </ul>
